GG-11732 Externalizeble objects

+++ modules/core/src/main/java/org/apache/ignite/internal/visor/cache/VisorCache.java
@@ -343,7 +343,7 @@ public class VisorCache extends VisorDataTransferObject {
         mode = CacheMode.fromOrdinal(in.readByte());
         memorySize = in.readLong();
         indexesSize = in.readLong();
-        size = in.readInt();
+        size = in.readLong();
         nearSize = in.readInt();
         primarySize = in.readLong();
         backupSize = in.readLong();
